+++ server/api/controllers/base-custom-field-groups/delete.js
@@ -45,6 +45,9 @@
 const { idInput } = require('../../../utils/inputs');
 
 const Errors = {
+  NOT_ENOUGH_RIGHTS: {
+    notEnoughRights: 'Not enough rights',
+  },
   BASE_CUSTOM_FIELD_GROUP_NOT_FOUND: {
     baseCustomFieldGroupNotFound: 'Base custom field group not found',
   },
@@ -59,6 +62,9 @@ module.exports = {
   },
 
   exits: {
+    notEnoughRights: {
+      responseType: 'forbidden',
+    },
     baseCustomFieldGroupNotFound: {
       responseType: 'notFound',
     },
@@ -78,6 +84,10 @@ module.exports = {
 
     if (!isProjectManager) {
       throw Errors.BASE_CUSTOM_FIELD_GROUP_NOT_FOUND; // Forbidden
+    }
+
+    if (project.isArchived) {
+      throw Errors.NOT_ENOUGH_RIGHTS;
     }
 
     baseCustomFieldGroup = await sails.helpers.baseCustomFieldGroups.deleteOne.with({
